+++ Changes.md
@@ -1,5 +1,7 @@
 ## WIP  TBD
 
+ * Add the `--extract-field/-X` option to name which fields are extracted from
+   the JSON output and displayed inline (default: `error,stacktrace`).
  * Hide null values from extra fields section.
  * Add the `--show-null` option to show null values in extra fields section.